Move buildinfo route into coderd.go
