--- coderd/buildinfo_test.go
@@ -1,19 +1,0 @@
-package coderd_test
-
-import (
-	"context"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	"github.com/coder/coder/cli/buildinfo"
-	"github.com/coder/coder/coderd/coderdtest"
-)
-
-func TestBuildInfo(t *testing.T) {
-	t.Parallel()
-	client := coderdtest.New(t, nil)
-	buildInfo, err := client.BuildInfo(context.Background())
-	require.NoError(t, err)
-	require.Equal(t, buildinfo.Version(), buildInfo.Version, "version")
-}
