+++ coderd/coderd_test.go
@@ -1,11 +1,25 @@
 package coderd_test
 
 import (
+	"context"
 	"testing"
 
 	"go.uber.org/goleak"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/coder/coder/cli/buildinfo"
+	"github.com/coder/coder/coderd/coderdtest"
 )
 
 func TestMain(m *testing.M) {
 	goleak.VerifyTestMain(m)
+}
+
+func TestBuildInfo(t *testing.T) {
+	t.Parallel()
+	client := coderdtest.New(t, nil)
+	buildInfo, err := client.BuildInfo(context.Background())
+	require.NoError(t, err)
+	require.Equal(t, buildinfo.Version(), buildInfo.Version, "version")
 }
